Clean up example contracts code
