+++ services/processor/native/sanitizer/test/usecases/create_channel.go
@@ -10,17 +10,11 @@ import (
 var PUBLIC = sdk.Export(add)
 var SYSTEM = sdk.Export(_init)
 
-var COUNTER_KEY = []byte("count")
-
 func _init() {
-	state.WriteUint64(COUNTER_KEY, 0)
 }
 
 func add(amount uint64) {
 	c := make(chan bool)
-
-	count := state.ReadUint64(COUNTER_KEY)
-	count += amount
-	state.WriteUint64(COUNTER_KEY, count)
+	c <- true
 }
 `
