+++ services/processor/native/sanitizer/test/usecases/send_to_channel.go
@@ -11,17 +11,10 @@ import (
 var PUBLIC = sdk.Export(add)
 var SYSTEM = sdk.Export(_init)
 
-var COUNTER_KEY = []byte("count")
-
 func _init() {
-	state.WriteUint64(COUNTER_KEY, 0)
 }
 
 func add(amount uint64) {
 	<- time.After(1*time.Minute)
-
-	count := state.ReadUint64(COUNTER_KEY)
-	count += amount
-	state.WriteUint64(COUNTER_KEY, count)
 }
 `
