+++ services/processor/native/sanitizer/test/usecases/use_goroutine.go
@@ -10,10 +10,7 @@ import (
 var PUBLIC = sdk.Export(add)
 var SYSTEM = sdk.Export(_init)
 
-var COUNTER_KEY = []byte("count")
-
 func _init() {
-	state.WriteUint64(COUNTER_KEY, 0)
 }
 
 func add(amount uint64) {
@@ -21,9 +18,5 @@ func add(amount uint64) {
 	go func() {
 		i = 10
 	}()
-
-	count := state.ReadUint64(COUNTER_KEY)
-	count += amount
-	state.WriteUint64(COUNTER_KEY, count)
 }
 `
